Add a stack view and basic navigation
