+++ src/App.tsx
@@ -1,51 +1,29 @@
 import React from 'react';
 import {
-  SafeAreaView,
   StyleSheet,
-  ScrollView,
-  Text,
-  Button,
   StatusBar,
-  View
 } from 'react-native';
-import { add } from './actions/countActions';
-import { useDispatch, useSelector } from 'react-redux';
+import 'react-native-gesture-handler';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
+import HomeContainer from './containers/home'
+
+const Stack = createStackNavigator();
 
 const App: React.FC = () => {
-  const dispatch = useDispatch();
-  const count = useSelector((state) => (state as any).count.count);
-  console.log('count = ', count);
-
-  const loginButtonStyles = {
-    justifyContent: 'center'
-  }
-
   return (
     <>
       <StatusBar barStyle="dark-content" />
-      <SafeAreaView style={styles.container}>
-        <View style={styles.authContainer}>
-          <Button
-            title="Google Login"
-            onPress={() => {
-              dispatch(add(1));
-            }}
+      <NavigationContainer>
+        <Stack.Navigator>
+          <Stack.Screen
+            name="Home"
+            component={HomeContainer}
+            options={{ title: 'Welcome' }}
           />
-        </View>
-        <ScrollView
-          contentInsetAdjustmentBehavior="automatic"
-          style={styles.booksContainer}
-        >
-          <Text>Hello world</Text>
-          <Text>{count}</Text>
-          <Button
-            title="Press Me"
-            onPress={() => {
-              dispatch(add(1));
-            }}
-          />
-        </ScrollView>
-      </SafeAreaView>
+          {/* <Stack.Screen name="Profile" component={ProfileScreen} /> */}
+        </Stack.Navigator>
+      </NavigationContainer>
     </>
   );
 };
